{"schema":"decodex/commit/1","summary":"Extract frozen brush model owner","authority":"manual"}
